refactor: extract drag-n-drop into hook
